+++ src/starkbank/boleto_holmes.clj
@@ -116,7 +116,7 @@
   "Receive a single BoletoHolmes map previously created by the Stark Bank API by passing its id
 
   ## Parameters (required):
-    - `id` [string]: map unique id. ex: \"5656565656565656\"
+    - `:id` [string]: map unique id. ex: \"5656565656565656\"
 
   ## Options:
     - `:user` [Project or Organization]: Project or Organization map returned from starkbank.user/project or starkbank.user/organization. Only necessary if starkbank.settings/user has not been set.
@@ -176,7 +176,7 @@
   "Receive a single Log map previously created by the Stark Bank API by passing its id
 
   ## Parameters (required):
-    - `id` [string]: map unique id. ex: \"5656565656565656\"
+    - `:id` [string]: map unique id. ex: \"5656565656565656\"
 
   ## Options:
     - `:user` [Project or Organization]: Project or Organization map returned from starkbank.user/project or starkbank.user/organization. Only necessary if starkbank.settings/user has not been set.