+++ CHANGELOG.md
@@ -15,9 +15,10 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 ## [Unreleased]
 ### Added
 - "payment" account type for Pix related resources
-- missing parameters to Boleto, BrcodePayment, DictKey, Event, Transfer and Workspace resources
-- workspace/update to allow parameter updatesa
+- missing parameters to boleto, brcode-payment, dict-key, event, transfer and workspace resources
+- workspace/update to allow parameter updates
 - event.attempt sub-resource to allow retrieval of information on failed webhook event delivery attempts
+- pdf method for retrieving PDF receipts from reversed invoice logs
 
 ## [2.3.4] - 2021-03-04
 ### Added